pyproject.toml
[project]
name = "whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-whisper",
    "psutil>=7.2.1",
    "pyside6>=6.10.1",
    "sounddevice>=0.5.3",
    "torch>=2.9.1",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
